Implemento o GET

+++ src/main/resources/db/migration/V1__create-table-produto.sql
@@ -3,7 +3,7 @@ create table produto(
     nome varchar(100) not null,
     peso varchar(10) not null,
     tipo varchar(20) not null,
-    valor float not null,
+    valor int not null,
     validade varchar(50) not null,
     
     primary key(id)
